requirements.txt
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
duckduckgo-search